Isolate teaching lesson flow
