+++ angular/src/app/playground/playground.component.html
@@ -10,7 +10,7 @@
 
     <!--Popup window with options for the documents-->
     <div class="menu" *ngIf="this.fabric.has_selection()">
-      <div class="br-icon" *ngIf="!this.lesson_mode" (click)="this.fabric.delete_selected()">
+      <div class="br-icon" *ngIf="!this.teaching.lesson_mode" (click)="this.fabric.delete_selected()">
         <mat-icon class="menu-icon">backspace</mat-icon> delete objects
       </div>
       <div class="br-icon" (click)="this.request_commentary()">
@@ -19,7 +19,7 @@
     </div>
 
     <button
-      *ngIf="this.lesson_commentary_button_visible"
+      *ngIf="this.teaching.lesson_mode && this.lesson_commentary_button_visible"
       mat-mini-fab
       class="lesson-fragment-commentary-button"
       [ngStyle]="this.lesson_commentary_button_style"
@@ -29,38 +29,38 @@
     </button>
 
     <!-- Lesson-mode banner -->
-    <div class="lesson-banner" *ngIf="this.lesson_mode">
+    <div class="lesson-banner" *ngIf="this.teaching.lesson_mode">
       <div class="lesson-text">
-        <div class="lesson-prompt">{{ this.current_prompt }}</div>
-        <div class="lesson-feedback" *ngIf="this.step_checked && this.current_step_score">
+        <div class="lesson-prompt">{{ this.teaching.current_prompt }}</div>
+        <div class="lesson-feedback" *ngIf="this.teaching.step_checked && this.teaching.current_step_score">
           <div class="lesson-score">
-            {{ this.current_step_score!.correct }}/{{ this.current_step_score!.total }} doel-fragmenten in de juiste zone.<span
-              *ngIf="this.current_step_score!.misplaced > 0">
-              {{ this.current_step_score!.misplaced }} afleider(s) verkeerd in de zone gesleept.</span>
+            {{ this.teaching.current_step_score!.correct }}/{{ this.teaching.current_step_score!.total }} doel-fragmenten in de juiste zone.<span
+              *ngIf="this.teaching.current_step_score!.misplaced > 0">
+              {{ this.teaching.current_step_score!.misplaced }} afleider(s) verkeerd in de zone gesleept.</span>
           </div>
         </div>
       </div>
-      <div class="lesson-step">Stap {{ this.current_step_index + 1 }} / {{ this.total_steps }}</div>
+      <div class="lesson-step">Stap {{ this.teaching.current_step_index + 1 }} / {{ this.teaching.total_steps }}</div>
       <button
-        *ngIf="!this.step_checked"
+        *ngIf="!this.teaching.step_checked"
         mat-stroked-button
         class="lesson-action"
-        (click)="this.check_step()">
+        (click)="this.teaching.check_step()">
         Controleer
       </button>
       <button
-        *ngIf="this.step_checked"
+        *ngIf="this.teaching.step_checked"
         mat-stroked-button
         class="lesson-action"
-        (click)="this.next_step()">
-        {{ this.current_step_index + 1 < this.total_steps ? 'Volgende stap' : 'Afronden' }}
+        (click)="this.teaching.next_step()">
+        {{ this.teaching.current_step_index + 1 < this.teaching.total_steps ? 'Volgende stap' : 'Afronden' }}
       </button>
       <button mat-icon-button class="lesson-exit" (click)="this.exit_lesson()">
         <mat-icon>close</mat-icon>
       </button>
     </div>
 
-    <div id="playground-main-toolbar" class="playground-toolbar" *ngIf="!this.lesson_mode">
+    <div id="playground-main-toolbar" class="playground-toolbar" *ngIf="!this.teaching.lesson_mode">
       <!-- Hamburger menu -->
       <div id="playground-hamburger-menu" class="playground-tool-group">
         <button id="playground-hamburger-menu-button" mat-stroked-button class="navbar" [matMenuTriggerFor]="menu">
@@ -218,7 +218,7 @@
     <div class="menu-item-icon-wrapper"></div>
     Settings
   </button>
-  <button mat-menu-item (click)="this.start_lesson()">
+  <button mat-menu-item (click)="this.teaching.start_lesson()">
     <div class="menu-item-icon-wrapper">
       <mat-icon>school</mat-icon>
     </div>
